Further breaking down computeFunctions. Currently runs correctly.
