+++ src/go_melt/core/computeFunctions.py
@@ -4,13 +4,9 @@ import dill
 import jax
 import jax.numpy as jnp
 import numpy as np
-from jax.experimental import sparse
-from jax.numpy import multiply
-from pyevtk.hl import gridToVTK
 import sys
-from go_melt.utils.gaussian_quadrature import (
-    computeQuad2dFemShapeFunctions_jax,
-    computeQuad3dFemShapeFunctions_jax,
+from go_melt.utils.interpolation_functions import (
+    interpolatePoints,
 )
 
 
@@ -77,54 +73,6 @@ def save_object(obj, filename):
         dill.dump(obj, outp, dill.HIGHEST_PROTOCOL)
 
 
-def getStaticNodesAndElements(L):
-    """
-    Extract static element and node counts from a list of level dictionaries.
-
-    Parameters:
-    L (list): A list of level dictionaries (e.g., from SetupLevels),
-              where each dictionary contains 'ne' (number of elements)
-              and 'nn' (number of nodes) as JAX arrays.
-
-    Returns:
-    tuple: A tuple containing:
-        - L[2]["ne"] as int
-        - L[3]["ne"] as int
-        - L[1]["nn"] as int
-        - L[2]["nn"] as int
-        - L[3]["nn"] as int
-    """
-    return (
-        L[2]["ne"].tolist(),
-        L[3]["ne"].tolist(),
-        L[1]["nn"].tolist(),
-        L[2]["nn"].tolist(),
-        L[3]["nn"].tolist(),
-    )
-
-
-def getStaticSubcycle(N):
-    """
-    Extract and compute subcycle values from the non-mesh configuration.
-
-    Parameters:
-    N (dict): Dictionary containing subcycle numbers for Level 2 and Level 3.
-
-    Returns:
-    tuple: A tuple containing:
-        - N2 (int): Subcycle count for Level 2
-        - N3 (int): Subcycle count for Level 3
-        - N23 (int): Total subcycles (N2 * N3)
-        - fN2 (float): N2 as float
-        - fN3 (float): N3 as float
-        - fN23 (float): N23 as float
-    """
-    N2, N3 = N["subcycle_num_L2"], N["subcycle_num_L3"]
-    N23 = N2 * N3
-    fN2, fN3, fN23 = float(N2), float(N3), float(N23)
-    return (N2, N3, N23, fN2, fN3, fN23)
-
-
 def calcStaticTmpNodesAndElements(L, v):
     """
     Calculate temporary element and node counts in Level 1 up to a given z-value.
@@ -150,48 +98,6 @@ def calcStaticTmpNodesAndElements(L, v):
     return (tmp_ne, tmp_nn)
 
 
-def getBCindices(x):
-    """
-    Compute boundary condition indices for a structured 3D mesh.
-
-    Parameters:
-    x (object): An object with attributes:
-        - nodes (list of int): Number of nodes in x, y, z directions.
-        - nn (int): Total number of nodes.
-
-    Returns:
-    list of jnp.ndarray: A list containing indices for:
-        [west, east, south, north, bottom, top] boundaries.
-    """
-    nx, ny, nz = x.nodes[0], x.nodes[1], x.nodes[2]
-    nn = x.nn
-
-    # Bottom face (z = 0)
-    bidx = jnp.arange(0, nx * ny)
-
-    # Top face (z = nz - 1)
-    tidx = jnp.arange(nx * ny * (nz - 1), nn)
-
-    # West face (x = 0)
-    widx = jnp.arange(0, nn, nx)
-
-    # East face (x = nx - 1)
-    eidx = jnp.arange(nx - 1, nn, nx)
-
-    # South face (y = 0)
-    sidx = jnp.arange(0, nx)[:, None] + (nx * ny * jnp.arange(0, nz))[None, :]
-    sidx = sidx.reshape(-1)
-
-    # North face (y = ny - 1)
-    nidx = (
-        jnp.arange(nx * (ny - 1), nx * ny)[:, None]
-        + (nx * ny * jnp.arange(0, nz))[None, :]
-    )
-    nidx = nidx.reshape(-1)
-
-    return [widx, eidx, sidx, nidx, bidx, tidx]
-
-
 def getSubstrateNodes(Levels):
     """
     Calculate the number of substrate nodes (z ≤ 0) for Levels 1 to 3.
@@ -210,116 +116,6 @@ def getSubstrateNodes(Levels):
         for L in Levels[:4]
     ]
     return tuple(substrate)
-
-
-@partial(jax.jit, static_argnames=["ne", "nn"])
-def solveMatrixFreeFE(Level, nn, ne, k, rhocp, dt, T, Fc, Corr):
-    """
-    Perform an explicit matrix-free finite element thermal solve.
-
-    This function computes the temperature update for a single timestep
-    using a matrix-free approach, avoiding global matrix assembly.
-
-    Parameters:
-    Level (dict): Mesh level containing connectivity and geometry.
-    nn (int): Total number of nodes.
-    ne (int): Total number of elements.
-    k (array): Thermal conductivity at each node.
-    rhocp (array): Volumetric density times heat capacity at each node.
-    dt (float): Time step size.
-    T (array): Temperature at the previous timestep.
-    Fc (array): Integrated right-hand side (e.g., heat source + surface BC).
-    Corr (array): Integrated T' correction terms.
-
-    Returns:
-    array: Updated temperature field for the next timestep.
-    """
-    nen = jnp.size(Level["connect"][0], 1)  # Nodes per element
-    ndim = 3  # 3D problem
-
-    coords = getSampleCoords(Level)
-    N, dNdx, wq = computeQuad3dFemShapeFunctions_jax(coords)
-    wq = wq[0][0]  # Quadrature weight
-
-    # Precompute shape function matrices
-    NTN = jnp.matmul(N.T, N)
-    BTB = jnp.zeros((nen, nen))
-    for idim in range(ndim):
-        BTB += jnp.matmul(dNdx[:, :, idim].T, dNdx[:, :, idim])
-
-    def calcVal(i):
-        _, _, _, idx = convert2XYZ(
-            i,
-            Level["elements"][0],
-            Level["elements"][1],
-            Level["nodes"][0],
-            Level["nodes"][1],
-        )
-        kvec = k[idx].mean()
-        mvec = rhocp[idx].mean() / dt
-
-        Me = jnp.sum(NTN * wq * mvec, axis=0)
-        Ke = BTB * kvec * wq
-        LHSe = jnp.diag(Me) - Ke
-
-        return jnp.matmul(LHSe, T[idx]), Me, idx
-
-    # Vectorized element-wise computation
-    vcalcVal = jax.vmap(calcVal)
-    aT, aMe, aidx = vcalcVal(jnp.arange(ne))
-
-    # Assemble global temperature and mass contributions
-    newT = jnp.bincount(aidx.reshape(-1), aT.reshape(-1), length=nn)
-    newM = jnp.bincount(aidx.reshape(-1), aMe.reshape(-1), length=nn)
-
-    return (newT + Fc + Corr) / newM
-
-
-@jax.jit
-def convert2XYZ(i, ne_x, ne_y, nn_x, nn_y):
-    """
-    Compute local element indices and global node connectivity in 3D.
-
-    Parameters:
-    i (int): Element index (flattened).
-    ne_x (int): Number of elements in the x-direction.
-    ne_y (int): Number of elements in the y-direction.
-    nn_x (int): Number of nodes in the x-direction.
-    nn_y (int): Number of nodes in the y-direction.
-
-    Returns:
-    tuple:
-        ix (int): Element index in x-direction.
-        iy (int): Element index in y-direction.
-        iz (int): Element index in z-direction.
-        idx (jnp.ndarray): Global node indices for the 8-node hexahedral element.
-    """
-    ne_xy = ne_x * ne_y
-    nn_xy = nn_x * nn_y
-
-    iz = i // ne_xy
-    iy = (i // ne_x) - iz * ne_y
-    ix = i % ne_x
-
-    base = ix + iy * nn_x + iz * nn_xy
-    dx = 1
-    dy = nn_x
-    dz = nn_xy
-
-    idx = jnp.array(
-        [
-            base,
-            base + dx,
-            base + dx + dy,
-            base + dy,
-            base + dz,
-            base + dx + dz,
-            base + dx + dy + dz,
-            base + dy + dz,
-        ]
-    )
-
-    return ix, iy, iz, idx
 
 
 def getCoarseNodesInFineRegion(xnf, xnc):
@@ -392,601 +188,6 @@ def getCoarseNodesInLargeFineRegion(xnc, xnf):
     return overlap
 
 
-@jax.jit
-def interpolatePointsMatrix(Level, node_coords_new):
-    """
-    Compute interpolation shape functions and node indices for mapping
-    values from a coarse mesh to a new set of coordinates.
-
-    Parameters:
-    Level (dict): Contains mesh information:
-                  - "node_coords": list of 1D arrays for x, y, z coordinates.
-                  - "connect": list of connectivity arrays in x, y, z.
-    node_coords_new (list): New nodal coordinates [x_new, y_new, z_new],
-                            each as a 2D array.
-
-    Returns:
-    list: [_Nc, _node]
-          _Nc (array): Shape function values for interpolation.
-          _node (array): Indices of coarse nodes contributing to each point.
-    """
-    ne_x = Level["connect"][0].shape[0]
-    ne_y = Level["connect"][1].shape[0]
-    ne_z = Level["connect"][2].shape[0]
-    nn_x, nn_y = ne_x + 1, ne_y + 1
-
-    nn_xn = len(node_coords_new[0])
-    nn_yn = len(node_coords_new[1])
-    nn_zn = len(node_coords_new[2])
-    total_points = nn_xn * nn_yn * nn_zn
-
-    h_x = Level["node_coords"][0][1] - Level["node_coords"][0][0]
-    h_y = Level["node_coords"][1][1] - Level["node_coords"][1][0]
-    h_z = Level["node_coords"][2][1] - Level["node_coords"][2][0]
-
-    def stepInterpolatePoints(ielt):
-        izn, rem = jnp.divmod(ielt, nn_xn * nn_yn)
-        iyn, ixn = jnp.divmod(rem, nn_xn)
-
-        _x = node_coords_new[0][ixn]
-        _y = node_coords_new[1][iyn]
-        _z = node_coords_new[2][izn]
-
-        # Determine coarse element indices
-        ielt_x = jnp.clip(
-            jnp.floor((_x - Level["node_coords"][0][0]) / h_x).astype(int), 0, ne_x - 1
-        )
-        ielt_y = jnp.clip(
-            jnp.floor((_y - Level["node_coords"][1][0]) / h_y).astype(int), 0, ne_y - 1
-        )
-        ielt_z = jnp.clip(
-            jnp.floor((_z - Level["node_coords"][2][0]) / h_z).astype(int), 0, ne_z - 1
-        )
-
-        # Get node indices
-        nodex = Level["connect"][0][ielt_x, :]
-        nodey = Level["connect"][1][ielt_y, :]
-        nodez = Level["connect"][2][ielt_z, :]
-        node = nodex + nodey * nn_x + nodez * (nn_x * nn_y)
-
-        # Get coordinates of the coarse element nodes
-        xx = Level["node_coords"][0][nodex]
-        yy = Level["node_coords"][1][nodey]
-        zz = Level["node_coords"][2][nodez]
-
-        # Bounding box corners
-        xc0, xc1 = xx[0], xx[1]
-        yc0, yc3 = yy[0], yy[3]
-        zc0, zc5 = zz[0], zz[5]
-
-        # Compute shape functions
-        Nc = compute3DN(
-            [_x, _y, _z], [xc0, xc1], [yc0, yc3], [zc0, zc5], [h_x, h_y, h_z]
-        )
-
-        # Clip and mask invalid values
-        valid = jnp.logical_and((Nc >= -1e-2).all(), (Nc <= 1 + 1e-2).all())
-        Nc = jax.lax.select(valid, jnp.clip(Nc, 0.0, 1.0), jnp.zeros_like(Nc))
-
-        return Nc, node
-
-    _Nc, _node = jax.vmap(stepInterpolatePoints)(jnp.arange(total_points))
-    return [_Nc, _node]
-
-
-@jax.jit
-def interpolate_w_matrix(C2F, T):
-    """
-    Interpolate a solution field to new nodal coordinates using shape functions.
-
-    This function applies precomputed shape functions and node indices
-    (from `interpolatePointsMatrix`) to interpolate the solution `T`
-    from a source mesh to a new set of points.
-
-    Parameters:
-    C2F (tuple): Interpolation data.
-                 - C2F[0]: Shape function weights (array of shape [n_new, n_basis]).
-                 - C2F[1]: Indices of source nodes (array of shape [n_new, n_basis]).
-    T (array): Source solution values at coarse nodes.
-
-    Returns:
-    array: Interpolated solution at new nodal coordinates.
-    """
-    return multiply(C2F[0], T[C2F[1]]).sum(axis=1)
-
-
-@jax.jit
-def interpolatePoints(Level, u, node_coords_new):
-    """
-    Interpolate a scalar field from a structured mesh to new coordinates.
-
-    This function evaluates shape functions at new nodal coordinates and
-    uses them to interpolate values from the source field `u` defined on
-    Level["node_coords"] and Level["connect"].
-
-    Parameters:
-    Level (dict): Contains mesh information:
-                  - "node_coords": list of 1D arrays for x, y, z coordinates.
-                  - "connect": list of connectivity arrays in x, y, z.
-    u (array): Field values at the source mesh nodes.
-    node_coords_new (list): New nodal coordinates [x_new, y_new, z_new].
-
-    Returns:
-    array: Interpolated values at the new nodal coordinates.
-    """
-    ne_x = Level["connect"][0].shape[0]
-    ne_y = Level["connect"][1].shape[0]
-    ne_z = Level["connect"][2].shape[0]
-    nn_x, nn_y = ne_x + 1, ne_y + 1
-
-    nn_xn = len(node_coords_new[0])
-    nn_yn = len(node_coords_new[1])
-    nn_zn = len(node_coords_new[2])
-    total_points = nn_xn * nn_yn * nn_zn
-
-    h_x = Level["node_coords"][0][1] - Level["node_coords"][0][0]
-    h_y = Level["node_coords"][1][1] - Level["node_coords"][1][0]
-    h_z = Level["node_coords"][2][1] - Level["node_coords"][2][0]
-
-    def stepInterpolatePoints(ielt):
-        izn, rem = jnp.divmod(ielt, nn_xn * nn_yn)
-        iyn, ixn = jnp.divmod(rem, nn_xn)
-
-        _x = node_coords_new[0][ixn]
-        _y = node_coords_new[1][iyn]
-        _z = node_coords_new[2][izn]
-
-        # Determine coarse element indices
-        ielt_x = jnp.clip(
-            jnp.floor((_x - Level["node_coords"][0][0]) / h_x).astype(int), 0, ne_x - 1
-        )
-        ielt_y = jnp.clip(
-            jnp.floor((_y - Level["node_coords"][1][0]) / h_y).astype(int), 0, ne_y - 1
-        )
-        ielt_z = jnp.clip(
-            jnp.floor((_z - Level["node_coords"][2][0]) / h_z).astype(int), 0, ne_z - 1
-        )
-
-        # Get node indices
-        nodex = Level["connect"][0][ielt_x, :]
-        nodey = Level["connect"][1][ielt_y, :]
-        nodez = Level["connect"][2][ielt_z, :]
-        node = nodex + nodey * nn_x + nodez * (nn_x * nn_y)
-
-        # Get coordinates of the coarse element nodes
-        xx = Level["node_coords"][0][nodex]
-        yy = Level["node_coords"][1][nodey]
-        zz = Level["node_coords"][2][nodez]
-
-        # Bounding box corners
-        xc0, xc1 = xx[0], xx[1]
-        yc0, yc3 = yy[0], yy[3]
-        zc0, zc5 = zz[0], zz[5]
-
-        # Compute shape functions
-        Nc = compute3DN(
-            [_x, _y, _z], [xc0, xc1], [yc0, yc3], [zc0, zc5], [h_x, h_y, h_z]
-        )
-
-        # Clip and mask invalid values
-        valid = jnp.logical_and((Nc >= -1e-2).all(), (Nc <= 1 + 1e-2).all())
-        Nc = jax.lax.select(valid, jnp.clip(Nc, 0.0, 1.0), jnp.zeros_like(Nc))
-
-        return Nc @ u[node]
-
-    return jax.vmap(stepInterpolatePoints)(jnp.arange(total_points))
-
-
-@jax.jit
-def computeCoarseFineShapeFunctions(Coarse, Fine):
-    """
-    Compute coarse shape functions and their derivatives at fine-scale
-    quadrature points, and return a sparse projection matrix.
-
-    Parameters:
-    Coarse (dict): Coarse mesh data.
-                   - "node_coords": list of 1D arrays for x, y, z coordinates.
-                   - "connect": list of connectivity arrays in x, y, z.
-    Fine (dict): Fine mesh data.
-                 - "node_coords": list of 1D arrays for x, y, z coordinates.
-                 - "connect": list of connectivity arrays in x, y, z.
-
-    Returns:
-    Nc (array): Coarse shape functions at fine quadrature points,
-                shape (n_fine_elem, 8, 8).
-    dNcdx, dNcdy, dNcdz (arrays): Derivatives of coarse shape functions,
-                                  each of shape (n_fine_elem, 8, 8).
-    test (BCOO): Sparse projection matrix from coarse nodes to fine quadrature.
-    """
-    # Mesh sizes
-    nec_x, nec_y, nec_z = [Coarse["connect"][i].shape[0] for i in range(3)]
-    nnc_x, nnc_y, nnc_z = [Coarse["node_coords"][i].shape[0] for i in range(3)]
-    nnc = nnc_x * nnc_y * nnc_z
-    nef_x, nef_y, nef_z = [Fine["connect"][i].shape[0] for i in range(3)]
-    nef = nef_x * nef_y * nef_z
-    nnf_x, nnf_y = [Fine["node_coords"][i].shape[0] for i in range(2)]
-
-    # Coarse mesh spacing
-    hc_x = Coarse["node_coords"][0][1] - Coarse["node_coords"][0][0]
-    hc_y = Coarse["node_coords"][1][1] - Coarse["node_coords"][1][0]
-    hc_z = Coarse["node_coords"][2][1] - Coarse["node_coords"][2][0]
-    hc_xyz = hc_x * hc_y * hc_z
-    xminc_x, xminc_y, xminc_z = [Coarse["node_coords"][i][0] for i in range(3)]
-
-    # Reference shape functions for fine elements
-    coords = jnp.stack(
-        [
-            Fine["node_coords"][0][Fine["connect"][0][0, :]],
-            Fine["node_coords"][1][Fine["connect"][1][0, :]],
-            Fine["node_coords"][2][Fine["connect"][2][0, :]],
-        ],
-        axis=1,
-    )
-    Nf, _, _ = computeQuad3dFemShapeFunctions_jax(coords)
-
-    def stepComputeCoarseFineTerm(ieltf):
-        ix, iy, iz, _ = convert2XYZ(ieltf, nef_x, nef_y, nnf_x, nnf_y)
-        coords_x = Fine["node_coords"][0][Fine["connect"][0][ix, :]].reshape(-1, 1)
-        coords_y = Fine["node_coords"][1][Fine["connect"][1][iy, :]].reshape(-1, 1)
-        coords_z = Fine["node_coords"][2][Fine["connect"][2][iz, :]].reshape(-1, 1)
-
-        x = (Nf @ coords_x).reshape(-1)
-        y = (Nf @ coords_y).reshape(-1)
-        z = (Nf @ coords_z).reshape(-1)
-
-        # Determine coarse element indices
-        ieltc_x = jnp.clip(jnp.floor((x - xminc_x) / hc_x).astype(int), 0, nec_x - 1)
-        ieltc_y = jnp.clip(jnp.floor((y - xminc_y) / hc_y).astype(int), 0, nec_y - 1)
-        ieltc_z = jnp.clip(jnp.floor((z - xminc_z) / hc_z).astype(int), 0, nec_z - 1)
-
-        def iqLoop(iq):
-            # Coarse element node indices
-            nodec_x = Coarse["connect"][0][ieltc_x[iq], :]
-            nodec_y = Coarse["connect"][1][ieltc_y[iq], :]
-            nodec_z = Coarse["connect"][2][ieltc_z[iq], :]
-            nodes = nodec_x + nodec_y * nnc_x + nodec_z * nnc_x * nnc_y
-
-            # Bounding box corners
-            xc0 = Coarse["node_coords"][0][nodec_x[0]]
-            xc1 = Coarse["node_coords"][0][nodec_x[1]]
-            yc0 = Coarse["node_coords"][1][nodec_y[0]]
-            yc3 = Coarse["node_coords"][1][nodec_y[3]]
-            zc0 = Coarse["node_coords"][2][nodec_z[0]]
-            zc5 = Coarse["node_coords"][2][nodec_z[5]]
-
-            _x, _y, _z = x[iq], y[iq], z[iq]
-
-            Nc = compute3DN(
-                [_x, _y, _z], [xc0, xc1], [yc0, yc3], [zc0, zc5], [hc_x, hc_y, hc_z]
-            )
-
-            dNcdx = (
-                jnp.array(
-                    [
-                        (-1 * (yc3 - _y) * (zc5 - _z)),
-                        (1 * (yc3 - _y) * (zc5 - _z)),
-                        (1 * (_y - yc0) * (zc5 - _z)),
-                        (-1 * (_y - yc0) * (zc5 - _z)),
-                        (-1 * (yc3 - _y) * (_z - zc0)),
-                        (1 * (yc3 - _y) * (_z - zc0)),
-                        (1 * (_y - yc0) * (_z - zc0)),
-                        (-1 * (_y - yc0) * (_z - zc0)),
-                    ]
-                )
-                / hc_xyz
-            )
-
-            dNcdy = (
-                jnp.array(
-                    [
-                        ((xc1 - _x) * -1 * (zc5 - _z)),
-                        ((_x - xc0) * -1 * (zc5 - _z)),
-                        ((_x - xc0) * 1 * (zc5 - _z)),
-                        ((xc1 - _x) * 1 * (zc5 - _z)),
-                        ((xc1 - _x) * -1 * (_z - zc0)),
-                        ((_x - xc0) * -1 * (_z - zc0)),
-                        ((_x - xc0) * 1 * (_z - zc0)),
-                        ((xc1 - _x) * 1 * (_z - zc0)),
-                    ]
-                )
-                / hc_xyz
-            )
-
-            dNcdz = (
-                jnp.array(
-                    [
-                        ((xc1 - _x) * (yc3 - _y) * -1),
-                        ((_x - xc0) * (yc3 - _y) * -1),
-                        ((_x - xc0) * (_y - yc0) * -1),
-                        ((xc1 - _x) * (_y - yc0) * -1),
-                        ((xc1 - _x) * (yc3 - _y) * 1),
-                        ((_x - xc0) * (yc3 - _y) * 1),
-                        ((_x - xc0) * (_y - yc0) * 1),
-                        ((xc1 - _x) * (_y - yc0) * 1),
-                    ]
-                )
-                / hc_xyz
-            )
-
-            return Nc, dNcdx, dNcdy, dNcdz, nodes
-
-        return jax.vmap(iqLoop)(jnp.arange(8))
-
-    Nc, dNcdx, dNcdy, dNcdz, _nodes = jax.vmap(stepComputeCoarseFineTerm)(
-        jnp.arange(nef)
-    )
-    _nodes = _nodes[:, 0, :]  # Only need one set of node indices per element
-
-    # Construct sparse projection matrix
-    indices = jnp.stack([_nodes.reshape(-1), jnp.arange(_nodes.size)], axis=1)
-    test = jax.experimental.sparse.BCOO(
-        (jnp.ones(_nodes.size), indices), shape=(nnc, _nodes.size)
-    )
-    return [Nc, [dNcdx, dNcdy, dNcdz], test]
-
-
-def compute3DN(q, x, y, z, h):
-    """
-    Compute trilinear shape functions for a hexahedral element at a given point.
-
-    Parameters:
-    q (list or array): Evaluation point [xq, yq, zq].
-    x (list): x-coordinates of the element's bounding box [x0, x1].
-    y (list): y-coordinates of the element's bounding box [y0, y1].
-    z (list): z-coordinates of the element's bounding box [z0, z1].
-    h (list): Element sizes in x, y, z directions [hx, hy, hz].
-
-    Returns:
-    array: Shape function values at point q, shape (8,).
-    """
-    inv_vol = 1.0 / (h[0] * h[1] * h[2])
-
-    N = (
-        jnp.array(
-            [
-                (x[1] - q[0]) * (y[1] - q[1]) * (z[1] - q[2]),
-                (q[0] - x[0]) * (y[1] - q[1]) * (z[1] - q[2]),
-                (q[0] - x[0]) * (q[1] - y[0]) * (z[1] - q[2]),
-                (x[1] - q[0]) * (q[1] - y[0]) * (z[1] - q[2]),
-                (x[1] - q[0]) * (y[1] - q[1]) * (q[2] - z[0]),
-                (q[0] - x[0]) * (y[1] - q[1]) * (q[2] - z[0]),
-                (q[0] - x[0]) * (q[1] - y[0]) * (q[2] - z[0]),
-                (x[1] - q[0]) * (q[1] - y[0]) * (q[2] - z[0]),
-            ]
-        )
-        * inv_vol
-    )
-
-    return N
-
-
-@jax.jit
-def computeCoarseTprimeMassTerm_jax(
-    Levels, Tprimef, Tprimem, L3rhocp, L2rhocp, dt, Shapes, Vcu, Vmu
-):
-    """
-    Compute coarse-scale mass terms from fine and medium-scale T' fields.
-
-    This function integrates the T' correction terms from Level 3 (fine)
-    and Level 2 (medium) and projects them to Level 1 and Level 2 using
-    precomputed shape transfer operators.
-
-    Parameters:
-    Levels (dict): Multilevel mesh hierarchy with keys [1, 2, 3].
-    Tprimef (array): Fine-scale T' field (Level 3).
-    Tprimem (array): Medium-scale T' field (Level 2).
-    L3rhocp (array): rhocp values at Level 3 nodes.
-    L2rhocp (array): rhocp values at Level 2 nodes.
-    dt (float): Time step size.
-    Shapes (list): Shape transfer operators between levels.
-                   Shapes[0]: L2 → L1
-                   Shapes[1]: L3 → L1
-                   Shapes[2]: L3 → L2
-    Vcu (array): Accumulator for Level 1 correction.
-    Vmu (array): Accumulator for Level 2 correction.
-
-    Returns:
-    Vcu (array): Updated Level 1 correction term.
-    Vmu (array): Updated Level 2 correction term.
-    """
-    # Subtract initial T' values
-    Tprimef_new = Tprimef - Levels[3]["Tprime0"]
-    Tprimem_new = Tprimem - Levels[2]["Tprime0"]
-
-    # Level 3 setup
-    nef_x, nef_y, nef_z = [Levels[3]["connect"][i].shape[0] for i in range(3)]
-    nef = nef_x * nef_y * nef_z
-    nnf_x = Levels[3]["node_coords"][0].shape[0]
-    nnf_y = Levels[3]["node_coords"][1].shape[0]
-
-    coordsf = getSampleCoords(Levels[3])
-    Nf, _, wqf = computeQuad3dFemShapeFunctions_jax(coordsf)
-
-    # Level 3
-    _, _, _, idxf = convert2XYZ(jnp.arange(nef), nef_x, nef_y, nnf_x, nnf_y)
-    _Tprimef = multiply(
-        Nf @ Tprimef_new[idxf], jnp.matmul(Nf, L3rhocp[idxf]).mean(axis=0)
-    )
-
-    _data1 = multiply(
-        multiply(-Shapes[1][0], _Tprimef.T[:, :, None]), (1 / dt) * wqf[None, None, :]
-    ).sum(axis=2)
-
-    _data2 = multiply(
-        multiply(-Shapes[2][0], _Tprimef.T[:, :, None]), (1 / dt) * wqf[None, None, :]
-    ).sum(axis=2)
-
-    # Level 2 setup
-    nem_x, nem_y, nem_z = [Levels[2]["connect"][i].shape[0] for i in range(3)]
-    nem = nem_x * nem_y * nem_z
-    nnm_x = Levels[2]["node_coords"][0].shape[0]
-    nnm_y = Levels[2]["node_coords"][1].shape[0]
-
-    coordsm = getSampleCoords(Levels[2])
-    Nm, _, wqm = computeQuad3dFemShapeFunctions_jax(coordsm)
-
-    _, _, _, idxm = convert2XYZ(jnp.arange(nem), nem_x, nem_y, nnm_x, nnm_y)
-    _Tprimem = multiply(
-        Nm @ Tprimem_new[idxm], jnp.matmul(Nm, L2rhocp[idxm]).mean(axis=0)
-    )
-
-    _data3 = multiply(
-        multiply(-Shapes[0][0], _Tprimem.T[:, :, None]), (1 / dt) * wqm[None, None, :]
-    ).sum(axis=2)
-
-    # Project to coarse levels
-    Vcu += Shapes[1][2] @ _data1.reshape(-1) + Shapes[0][2] @ _data3.reshape(-1)
-    Vmu += Shapes[2][2] @ _data2.reshape(-1)
-
-    return Vcu, Vmu
-
-
-@jax.jit
-def computeCoarseTprimeTerm_jax(Levels, L3k, L2k, Shapes):
-    """
-    Compute coarse-scale projection of T' gradient terms from fine and medium levels.
-
-    This function computes the gradient of the T' field at Level 3 and Level 2,
-    multiplies it by thermal conductivity, and projects the result to coarser
-    levels using precomputed shape transfer operators.
-
-    Parameters:
-    Levels (dict): Multilevel mesh hierarchy with keys [1, 2, 3].
-                   Each level contains "connect", "node_coords", and "Tprime0".
-    L3k (array): Thermal conductivity at Level 3 nodes.
-    L2k (array): Thermal conductivity at Level 2 nodes.
-    Shapes (list): Shape transfer operators between levels.
-                   Shapes[0]: L2 → L1
-                   Shapes[1]: L3 → L1
-                   Shapes[2]: L3 → L2
-
-    Returns:
-    Vcu (array): Coarse-scale correction term for Level 1.
-    Vmu (array): Coarse-scale correction term for Level 2.
-    """
-    # Level 3 setup
-    nef_x, nef_y, nef_z = [Levels[3]["connect"][i].shape[0] for i in range(3)]
-    nef = nef_x * nef_y * nef_z
-    nnf_x = Levels[3]["node_coords"][0].shape[0]
-    nnf_y = Levels[3]["node_coords"][1].shape[0]
-
-    coordsf = getSampleCoords(Levels[3])
-    Nf, dNdxf, wqf = computeQuad3dFemShapeFunctions_jax(coordsf)
-
-    _, _, _, idxf = convert2XYZ(jnp.arange(nef), nef_x, nef_y, nnf_x, nnf_y)
-    Tprimef = Levels[3]["Tprime0"][idxf]
-    L3kMean = jnp.matmul(Nf, L3k[idxf]).mean(axis=0)
-
-    dTprimefdx = multiply(L3kMean, dNdxf[:, :, 0] @ Tprimef)
-    dTprimefdy = multiply(L3kMean, dNdxf[:, :, 1] @ Tprimef)
-    dTprimefdz = multiply(L3kMean, dNdxf[:, :, 2] @ Tprimef)
-
-    # Project to Level 1 and Level 2
-    _data1 = sum(
-        [
-            multiply(
-                multiply(-Shapes[1][1][i], d.T[:, :, None]), wqf[None, None, :]
-            ).sum(axis=2)
-            for i, d in enumerate([dTprimefdx, dTprimefdy, dTprimefdz])
-        ]
-    )
-    _data2 = sum(
-        [
-            multiply(
-                multiply(-Shapes[2][1][i], d.T[:, :, None]), wqf[None, None, :]
-            ).sum(axis=2)
-            for i, d in enumerate([dTprimefdx, dTprimefdy, dTprimefdz])
-        ]
-    )
-
-    # Level 2 setup
-    nem_x, nem_y, nem_z = [Levels[2]["connect"][i].shape[0] for i in range(3)]
-    nem = nem_x * nem_y * nem_z
-    nnm_x = Levels[2]["node_coords"][0].shape[0]
-    nnm_y = Levels[2]["node_coords"][1].shape[0]
-
-    coordsm = getSampleCoords(Levels[2])
-    Nm, dNdxm, wqm = computeQuad3dFemShapeFunctions_jax(coordsm)
-
-    _, _, _, idxm = convert2XYZ(jnp.arange(nem), nem_x, nem_y, nnm_x, nnm_y)
-    Tprimem = Levels[2]["Tprime0"][idxm]
-    L2kMean = jnp.matmul(Nm, L2k[idxm]).mean(axis=0)
-
-    dTprimemdx = multiply(L2kMean, dNdxm[:, :, 0] @ Tprimem)
-    dTprimemdy = multiply(L2kMean, dNdxm[:, :, 1] @ Tprimem)
-    dTprimemdz = multiply(L2kMean, dNdxm[:, :, 2] @ Tprimem)
-
-    _data3 = sum(
-        [
-            multiply(
-                multiply(-Shapes[0][1][i], d.T[:, :, None]), wqm[None, None, :]
-            ).sum(axis=2)
-            for i, d in enumerate([dTprimemdx, dTprimemdy, dTprimemdz])
-        ]
-    )
-
-    # Final projection to coarse levels
-    Vcu = Shapes[1][2] @ _data1.reshape(-1) + Shapes[0][2] @ _data3.reshape(-1)
-    Vmu = Shapes[2][2] @ _data2.reshape(-1)
-
-    return Vcu, Vmu
-
-
-@jax.jit
-def assignBCs(RHS, Levels):
-    """
-    Apply Dirichlet boundary conditions to the right-hand side vector.
-
-    This function sets prescribed values at boundary nodes for Level 1
-    based on the boundary condition indices and values stored in the
-    Levels dictionary.
-
-    Parameters:
-    RHS (array): Right-hand side vector to be modified.
-    Levels (dict): Multilevel mesh hierarchy containing:
-                   - Levels[1]["BC"]: list of boundary node index arrays
-                     [x0, x1, y0, y1, z0]
-                   - Levels[1]["conditions"]: dict with keys "x", "y", "z"
-                     containing prescribed values.
-
-    Returns:
-    array: Modified RHS vector with boundary conditions applied.
-    """
-    _RHS = RHS
-    _RHS = _RHS.at[Levels[1]["BC"][2]].set(Levels[1]["conditions"]["y"][0])  # y-min
-    _RHS = _RHS.at[Levels[1]["BC"][3]].set(Levels[1]["conditions"]["y"][1])  # y-max
-    _RHS = _RHS.at[Levels[1]["BC"][0]].set(Levels[1]["conditions"]["x"][0])  # x-min
-    _RHS = _RHS.at[Levels[1]["BC"][1]].set(Levels[1]["conditions"]["x"][1])  # x-max
-    _RHS = _RHS.at[Levels[1]["BC"][4]].set(Levels[1]["conditions"]["z"][0])  # z-min
-
-    return _RHS
-
-
-@jax.jit
-def assignBCsFine(RHS, TfAll, BC):
-    """
-    Apply Dirichlet boundary conditions to the fine-level RHS vector.
-
-    This function sets the RHS values at boundary nodes using the
-    corresponding values from the full fine-scale solution `TfAll`.
-
-    Parameters:
-    RHS (array): Right-hand side vector to be modified.
-    TfAll (array): Full fine-scale solution field.
-    BC (list): List of boundary node index arrays [x0, x1, y0, y1, z0].
-
-    Returns:
-    array: Modified RHS vector with boundary values applied.
-    """
-    _RHS = RHS
-    _RHS = _RHS.at[BC[2]].set(TfAll[BC[2]])  # y-min
-    _RHS = _RHS.at[BC[3]].set(TfAll[BC[3]])  # y-max
-    _RHS = _RHS.at[BC[0]].set(TfAll[BC[0]])  # x-min
-    _RHS = _RHS.at[BC[1]].set(TfAll[BC[1]])  # x-max
-    _RHS = _RHS.at[BC[4]].set(TfAll[BC[4]])  # z-min
-    return _RHS
-
-
 @partial(jax.jit, static_argnames=["nn"])
 def bincount(N, D, nn):
     """
@@ -1036,182 +237,6 @@ def getOverlapRegion(node_coords, nx, ny):
     return _x + _y * nx + _z * nx * ny
 
 
-@jax.jit
-def jit_constrain_v(vtot, Level):
-    """
-    Constrain a 3D vector within the bounding box defined in the mesh level.
-
-    This function clips each component of the input vector `vtot` to lie
-    within the bounds specified in `Level["bounds"]`.
-
-    Parameters:
-    vtot (list): A list of 3 elements [vx, vy, vz] representing a 3D vector.
-    Level (dict): Contains bounding box limits under:
-                  - Level["bounds"]["ix"]: (xmin, xmax)
-                  - Level["bounds"]["iy"]: (ymin, ymax)
-                  - Level["bounds"]["iz"]: (zmin, zmax)
-
-    Returns:
-    list: Clipped 3D vector [vx_clipped, vy_clipped, vz_clipped].
-    """
-    vtot = [
-        jnp.clip(vtot[0], Level["bounds"]["ix"][0], Level["bounds"]["ix"][1]),
-        jnp.clip(vtot[1], Level["bounds"]["iy"][0], Level["bounds"]["iy"][1]),
-        jnp.clip(vtot[2], Level["bounds"]["iz"][0], Level["bounds"]["iz"][1]),
-    ]
-    return vtot
-
-
-@jax.jit
-def move_fine_mesh(node_coords, element_size, v):
-    """
-    Shift a structured fine mesh in space based on a displacement vector.
-
-    This function computes the new coordinates of a structured fine mesh
-    by translating it in the x, y, and z directions according to the
-    displacement vector `v`, scaled by the element size.
-
-    Parameters:
-    node_coords (list): Original node coordinates [x, y, z] as 1D arrays.
-    element_size (list): Element sizes [hx, hy, hz] in each direction.
-    v (array): Displacement vector [vx, vy, vz].
-
-    Returns:
-    list: New node coordinates [xnf_x, xnf_y, xnf_z] after translation.
-    list: Integer shift indices [vx_, vy_, vz_] used for the translation.
-    """
-    vx_ = (v[0] / element_size[0] + 1e-2).astype(int)
-    vy_ = (v[1] / element_size[1] + 1e-2).astype(int)
-    vz_ = (v[2] / element_size[2] + 1e-2).astype(int)
-
-    xnf_x = node_coords[0] + element_size[0] * vx_
-    xnf_y = node_coords[1] + element_size[1] * vy_
-    xnf_z = node_coords[2] + element_size[2] * vz_
-
-    return [xnf_x, xnf_y, xnf_z], [vx_, vy_, vz_]
-
-
-@jax.jit
-def update_overlap_nodes_coords(Level, vcon, element_size, ele_ratio):
-    """
-    Update the overlap node indices and coordinates based on a displacement vector.
-
-    This function shifts the overlap region of a mesh by updating both the
-    node indices and physical coordinates using the displacement vector `vcon`.
-
-    Parameters:
-    Level (dict): Mesh level containing original overlap node indices and coordinates.
-                  - "orig_overlap_nodes": list of original node index arrays [x, y, z].
-                  - "orig_overlap_coors": list of original coordinate arrays [x, y, z].
-    vcon (array): Displacement vector [vx, vy, vz].
-    element_size (array): Element sizes [hx, hy, hz].
-    ele_ratio (array): Ratio of coarse-to-fine elements in each direction.
-
-    Returns:
-    dict: Updated Level dictionary with new "overlapNodes" and "overlapCoords".
-    """
-    shift = [
-        (vcon[0] / element_size[0] + 1e-2).astype(int),
-        (vcon[1] / element_size[1] + 1e-2).astype(int),
-        (vcon[2] / element_size[2] + 1e-2).astype(int),
-    ]
-
-    Level["overlapNodes"] = [
-        Level["orig_overlap_nodes"][0] + ele_ratio[0] * shift[0],
-        Level["orig_overlap_nodes"][1] + ele_ratio[1] * shift[1],
-        Level["orig_overlap_nodes"][2] + ele_ratio[2] * shift[2],
-    ]
-
-    Level["overlapCoords"] = [
-        Level["orig_overlap_coors"][0] + element_size[0] * shift[0],
-        Level["orig_overlap_coors"][1] + element_size[1] * shift[1],
-        Level["orig_overlap_coors"][2] + element_size[2] * shift[2],
-    ]
-
-    return Level
-
-
-@jax.jit
-def update_overlap_nodes_coords_L1L2(Level, vcon, element_size, powder_layer):
-    """
-    Update overlap node indices and coordinates for Level 1/2 with powder layer offset.
-
-    This function shifts the overlap region of a mesh by updating both the
-    node indices and physical coordinates using the displacement vector `vcon`,
-    accounting for an additional powder layer in the z-direction.
-
-    Parameters:
-    Level (dict): Mesh level containing original overlap node indices and coordinates.
-                  - "orig_overlap_nodes": list of original node index arrays [x, y, z].
-                  - "orig_overlap_coors": list of original coordinate arrays [x, y, z].
-    vcon (array): Displacement vector [vx, vy, vz].
-    element_size (array): Element sizes [hx, hy, hz].
-    powder_layer (float): Additional offset in the z-direction.
-
-    Returns:
-    dict: Updated Level dictionary with new "overlapNodes" and "overlapCoords".
-    """
-    shift_x = (vcon[0] / element_size[0] + 1e-2).astype(int)
-    shift_y = (vcon[1] / element_size[1] + 1e-2).astype(int)
-    shift_z = (vcon[2] / element_size[2] + 1e-2).astype(int)
-    shift_z_p = powder_layer * (vcon[2] / powder_layer + 1e-2).astype(int)
-
-    Level["overlapNodes"] = [
-        Level["orig_overlap_nodes"][0] + shift_x,
-        Level["orig_overlap_nodes"][1] + shift_y,
-        Level["orig_overlap_nodes"][2] + shift_z,
-    ]
-
-    Level["overlapCoords"] = [
-        Level["orig_overlap_coors"][0] + element_size[0] * shift_x,
-        Level["orig_overlap_coors"][1] + element_size[1] * shift_y,
-        Level["orig_overlap_coors"][2] + shift_z_p,
-    ]
-
-    return Level
-
-
-@jax.jit
-def update_overlap_nodes_coords_L2(Level, vcon, element_size, ele_ratio):
-    """
-    Update overlap node indices and coordinates for Level 2 based on displacement.
-
-    This function shifts the overlap region of Level 2 by updating both the
-    node indices and physical coordinates using the displacement vector `vcon`
-    and the element-to-node ratio `ele_ratio`.
-
-    Parameters:
-    Level (dict): Mesh level containing original overlap node indices and coordinates.
-                - "orig_overlap_nodes_L2": list of original node index arrays [x, y, z].
-                - "orig_overlap_coors_L2": list of original coordinate arrays [x, y, z].
-    vcon (array): Displacement vector [vx, vy, vz].
-    element_size (array): Element sizes [hx, hy, hz].
-    ele_ratio (array): Ratio of coarse-to-fine elements in each direction.
-
-    Returns:
-    dict: Updated Level dictionary with new "overlapNodes_L2" and "overlapCoords_L2".
-    """
-    shift = [
-        (vcon[0] / element_size[0] + 1e-2).astype(int),
-        (vcon[1] / element_size[1] + 1e-2).astype(int),
-        (vcon[2] / element_size[2] + 1e-2).astype(int),
-    ]
-
-    Level["overlapNodes_L2"] = [
-        Level["orig_overlap_nodes_L2"][0] + ele_ratio[0] * shift[0],
-        Level["orig_overlap_nodes_L2"][1] + ele_ratio[1] * shift[1],
-        Level["orig_overlap_nodes_L2"][2] + ele_ratio[2] * shift[2],
-    ]
-
-    Level["overlapCoords_L2"] = [
-        Level["orig_overlap_coors_L2"][0] + element_size[0] * shift[0],
-        Level["orig_overlap_coors_L2"][1] + element_size[1] * shift[1],
-        Level["orig_overlap_coors_L2"][2] + element_size[2] * shift[2],
-    ]
-
-    return Level
-
-
 @partial(jax.jit, static_argnames=["_idx", "_val"])
 def substitute_Tbar(Tbar, _idx, _val):
     """
@@ -1250,36 +275,6 @@ def substitute_Tbar2(Tbar, _idx, _val):
     return Tbar.at[_idx].set(_val)
 
 
-def find_max_const(CoarseLevel, FinerLevel):
-    """
-    Compute the maximum number of elements the finer level domain can move
-    within the bounds of the coarser level domain in each direction.
-
-    This function calculates how many elements the finer mesh can shift
-    in the positive and negative directions (east/west, north/south, top/bottom)
-    without exceeding the bounds of the coarser mesh.
-
-    Parameters:
-    CoarseLevel (object): Object with `bounds` attribute containing:
-                          - bounds.x, bounds.y, bounds.z: tuples (min, max)
-    FinerLevel (object): Object with `bounds` attribute structured similarly.
-
-    Returns:
-    tuple: Three lists representing allowable movement in:
-           - x-direction: [west, east]
-           - y-direction: [south, north]
-           - z-direction: [bottom, top]
-    """
-    iE = CoarseLevel.bounds.x[1] - FinerLevel.bounds.x[1]  # Elements to east
-    iN = CoarseLevel.bounds.y[1] - FinerLevel.bounds.y[1]  # Elements to north
-    iT = CoarseLevel.bounds.z[1] - FinerLevel.bounds.z[1]  # Elements to top
-    iW = CoarseLevel.bounds.x[0] - FinerLevel.bounds.x[0]  # Elements to west
-    iS = CoarseLevel.bounds.y[0] - FinerLevel.bounds.y[0]  # Elements to south
-    iB = CoarseLevel.bounds.z[0] - FinerLevel.bounds.z[0]  # Elements to bottom
-
-    return [iW, iE], [iS, iN], [iB, iT]
-
-
 def calc_length_h(A):
     """
     Compute domain lengths and element sizes in each spatial direction.
@@ -1307,478 +302,6 @@ def calc_length_h(A):
     hz = Lz / A.elements[2]
 
     return [Lx, Ly, Lz], [hx, hy, hz]
-
-
-def saveResult(Level, save_str, record_lab, save_path, zoffset):
-    """
-    Save the temperature and state fields of a mesh level to a VTK file.
-
-    This function exports the structured grid data for visualization,
-    including temperature and material state fields, using the VTK format.
-
-    Parameters:
-    Level (dict): Dictionary containing mesh and field data:
-                  - "node_coords": [x, y, z] coordinate arrays
-                  - "T0": temperature field (flattened)
-                  - "S1": state field (flattened)
-                  - "nodes": [nx, ny, nz] number of nodes in each direction
-    save_str (str): Prefix for the output filename.
-    record_lab (int): Frame or timestep label for file naming.
-    save_path (str): Directory path to save the output file.
-    zoffset (float): Offset applied to z-coordinates for rendering purposes.
-
-    Returns:
-    None
-    """
-    # List coordinates in each direction for structured save
-    vtkcx = np.array(Level["node_coords"][0])
-    vtkcy = np.array(Level["node_coords"][1])
-    vtkcz = np.array(Level["node_coords"][2] - zoffset)
-
-    # Reshape the temperature and state fields for correct rendering
-    vtkT = np.array(
-        Level["T0"].reshape(Level["nodes"][2], Level["nodes"][1], Level["nodes"][0])
-    ).transpose((2, 1, 0))
-    vtkS = np.array(
-        Level["S1"].reshape(Level["nodes"][2], Level["nodes"][1], Level["nodes"][0])
-    ).transpose((2, 1, 0))
-
-    # Save a VTK rectilinear grid file
-    pointData = {"Temperature (K)": vtkT, "State (Powder/Solid)": vtkS}
-    vtkSave = f"{save_path}{save_str}{record_lab:08}"
-    gridToVTK(vtkSave, vtkcx, vtkcy, vtkcz, pointData=pointData)
-
-
-def saveFinalResult(Level, save_str, save_path, zoffset):
-    """
-    Save the final temperature and state fields of a mesh level to a VTK file.
-
-    This function exports the final structured grid data for visualization,
-    including temperature and material state fields, using the VTK format.
-
-    Parameters:
-    Level (dict): Dictionary containing mesh and field data:
-                  - "node_coords": [x, y, z] coordinate arrays
-                  - "T0": temperature field (flattened)
-                  - "S1": state field (flattened)
-                  - "nodes": [nx, ny, nz] number of nodes in each direction
-    save_str (str): Prefix for the output filename.
-    save_path (str): Directory path to save the output file.
-    zoffset (float): Offset applied to z-coordinates for rendering purposes.
-
-    Returns:
-    None
-    """
-    # List coordinates in each direction for structured save
-    vtkcx = np.array(Level["node_coords"][0])
-    vtkcy = np.array(Level["node_coords"][1])
-    vtkcz = np.array(Level["node_coords"][2] - zoffset)
-
-    # Reshape the temperature and state fields for correct rendering
-    vtkT = np.array(
-        Level["T0"].reshape(Level["nodes"][2], Level["nodes"][1], Level["nodes"][0])
-    ).transpose((2, 1, 0))
-    vtkS = np.array(
-        Level["S1"].reshape(Level["nodes"][2], Level["nodes"][1], Level["nodes"][0])
-    ).transpose((2, 1, 0))
-
-    # Save a VTK rectilinear grid file
-    pointData = {"Temperature (K)": vtkT, "State (Powder/Solid)": vtkS}
-    vtkSave = f"{save_path}{save_str}Final"
-    gridToVTK(vtkSave, vtkcx, vtkcy, vtkcz, pointData=pointData)
-
-
-def saveState(Level, save_str, record_lab, save_path, zoffset):
-    """
-    Save the current state field of a mesh level to a VTK file.
-
-    This function exports the structured grid data for visualization,
-    including only the material state field (e.g., powder or solid).
-
-    Parameters:
-    Level (dict): Dictionary containing mesh and field data:
-                  - "node_coords": [x, y, z] coordinate arrays
-                  - "S1": state field (flattened)
-                  - "nodes": [nx, ny, nz] number of nodes in each direction
-    save_str (str): Prefix for the output filename.
-    record_lab (int): Frame or timestep label for file naming.
-    save_path (str): Directory path to save the output file.
-    zoffset (float): Offset applied to z-coordinates for rendering purposes.
-
-    Returns:
-    None
-    """
-    # List coordinates in each direction for structured save
-    vtkcx = np.array(Level["node_coords"][0])
-    vtkcy = np.array(Level["node_coords"][1])
-    vtkcz = np.array(Level["node_coords"][2] - zoffset)
-
-    # Reshape the state field for correct rendering
-    vtkS = np.array(
-        Level["S1"].reshape(Level["nodes"][2], Level["nodes"][1], Level["nodes"][0])
-    ).transpose((2, 1, 0))
-
-    # Save a VTK rectilinear grid file
-    pointData = {"State (Powder/Solid)": vtkS}
-    vtkSave = f"{save_path}{save_str}{record_lab:08}"
-    gridToVTK(vtkSave, vtkcx, vtkcy, vtkcz, pointData=pointData)
-
-
-@jax.jit
-def getNewTprime(Fine, FineT0, CoarseT, Coarse, C2F):
-    """
-    Compute the fine-level temperature residual (T') and update the coarse-level
-    temperature.
-
-    This function interpolates the fine-level temperature onto the overlap region,
-    updates the coarse-level temperature at the overlap, and computes the temperature
-    difference (residual) between the fine level and the interpolated coarse level.
-
-    Parameters:
-    Fine (dict): Fine level mesh and metadata, including:
-                 - "overlapCoords": coordinates for interpolation
-                 - "overlapNodes": node indices for overlap region
-    FineT0 (array): Fine level temperature field.
-    CoarseT (array): Coarse level temperature field (to be updated).
-    Coarse (dict): Coarse level mesh and metadata, including:
-                   - "nodes": [nx, ny, nz] node counts
-    C2F (array): Coarse-to-fine interpolation matrix.
-
-    Returns:
-    tuple:
-        - Tprime (array): Fine-level temperature residual.
-        - CoarseT (array): Updated coarse-level temperature field.
-    """
-    # Interpolate fine temperature at overlap coordinates
-    _val = interpolatePoints(Fine, FineT0, Fine["overlapCoords"])
-
-    # Get flattened index of overlap region in coarse mesh
-    _idx = getOverlapRegion(
-        Fine["overlapNodes"], Coarse["nodes"][0], Coarse["nodes"][1]
-    )
-
-    # Update coarse temperature at overlap region
-    CoarseT = substitute_Tbar2(CoarseT, _idx, _val)
-
-    # Compute residual between fine and interpolated coarse temperature
-    Tprime = FineT0 - interpolate_w_matrix(C2F, CoarseT)
-
-    return Tprime, CoarseT
-
-
-@jax.jit
-def getBothNewTprimes(Levels, FineT, MesoT, M2F, CoarseT, C2M):
-    """
-    Compute temperature residuals at two nested levels in a multilevel hierarchy.
-
-    This function computes the temperature residuals (T') for both the fine and
-    meso levels by updating and interpolating across the hierarchy:
-    Fine → Meso → Coarse.
-
-    Parameters:
-    Levels (dict): Dictionary of mesh levels:
-                   - Levels[1]: Coarse
-                   - Levels[2]: Meso
-                   - Levels[3]: Fine
-    FineT (array): Fine level temperature field.
-    MesoT (array): Meso level temperature field.
-    M2F (array): Meso-to-fine interpolation matrix.
-    CoarseT (array): Coarse level temperature field.
-    C2M (array): Coarse-to-meso interpolation matrix.
-
-    Returns:
-    tuple:
-        - lTprime (array): Fine-level temperature residual.
-        - mTprime (array): Meso-level temperature residual.
-        - mT0 (array): Updated meso-level temperature field.
-        - uT0 (array): Updated coarse-level temperature field.
-    """
-    lTprime, mT0 = getNewTprime(Levels[3], FineT, MesoT, Levels[2], M2F)
-    mTprime, uT0 = getNewTprime(Levels[2], mT0, CoarseT, Levels[1], C2M)
-
-    return lTprime, mTprime, mT0, uT0
-
-
-@partial(jax.jit, static_argnames=["ne_nn", "tmp_ne_nn"])
-def computeSolutions(
-    Levels, ne_nn, tmp_ne_nn, LF, L1V, LInterp, Lk, Lrhocp, L2V, dt, properties
-):
-    """
-    Compute temperature solutions across three mesh levels using a matrix-free FEM solver.
-
-    This function performs a multiscale temperature update by solving the heat equation
-    at coarse, meso, and fine levels. It uses matrix-free finite element methods and
-    interpolates between levels to apply boundary conditions and source terms.
-
-    Parameters:
-    Levels (dict): Dictionary of mesh levels:
-                   - Levels[1]: Coarse
-                   - Levels[2]: Meso
-                   - Levels[3]: Fine
-    ne_nn (list): List of neighbor-element and node-element mappings for each level.
-    tmp_ne_nn (list): Temporary mappings and boundary indices for Level 1.
-    LF (dict): Source terms for each level.
-    L1V (array): Volume weights for Level 1.
-    LInterp (list): Interpolation matrices:
-                    - LInterp[0]: Level 1 → Level 2
-                    - LInterp[1]: Level 2 → Level 3
-    Lk (dict): Thermal conductivity for each level.
-    Lrhocp (dict): Volumetric heat capacity for each level.
-    L2V (array): Volume weights for Level 2.
-    dt (float): Time step size.
-    properties (dict): Material properties, including:
-                       - "T_amb": Ambient temperature for Dirichlet BCs.
-
-    Returns:
-    tuple:
-        - FinalL1 (array): Updated temperature field at Level 1.
-        - FinalL2 (array): Updated temperature field at Level 2.
-        - FinalL3 (array): Updated temperature field at Level 3.
-    """
-    # Solve coarse-level problem (Level 1)
-    L1T = solveMatrixFreeFE(
-        Levels[1],
-        ne_nn[2],
-        tmp_ne_nn[0],
-        Lk[1],
-        Lrhocp[1],
-        dt,
-        Levels[1]["T0"],
-        LF[1],
-        L1V,
-    )
-    L1T = substitute_Tbar(L1T, tmp_ne_nn[1], properties["T_amb"])
-    FinalL1 = assignBCs(L1T, Levels)
-
-    # Interpolate Level 1 solution to Level 2 for source term
-    TfAll = interpolate_w_matrix(LInterp[0], FinalL1)
-
-    # Solve meso-level problem (Level 2)
-    L2T = solveMatrixFreeFE(
-        Levels[2], ne_nn[3], ne_nn[0], Lk[2], Lrhocp[2], dt, Levels[2]["T0"], LF[2], L2V
-    )
-    FinalL2 = assignBCsFine(L2T, TfAll, Levels[2]["BC"])
-
-    # Interpolate Level 2 solution to Level 3 for boundary conditions
-    TfAll = interpolate_w_matrix(LInterp[1], FinalL2)
-
-    # Solve fine-level problem (Level 3)
-    FinalL3 = solveMatrixFreeFE(
-        Levels[3], ne_nn[4], ne_nn[1], Lk[3], Lrhocp[3], dt, Levels[3]["T0"], LF[3], 0
-    )
-    FinalL3 = assignBCsFine(FinalL3, TfAll, Levels[3]["BC"])
-
-    return FinalL1, FinalL2, FinalL3
-
-
-@partial(jax.jit, static_argnames=["ne", "nn"])
-def computeConvRadBC(Level, LevelT0, ne, nn, properties, F):
-    """
-    Compute Neumann boundary conditions on the top surface due to:
-      • Convection (Newton's law of cooling)
-      • Radiation (Stefan-Boltzmann law)
-      • Evaporation (empirical model)
-
-    The resulting heat flux is integrated using 2D quadrature and added
-    to the global force vector.
-
-    Parameters:
-    Level (dict): Mesh data with:
-                  - "node_coords": [x, y] coordinate arrays
-                  - "connect": [cx, cy] element connectivity arrays
-    LevelT0 (array): Temperature field at current time step.
-    ne (int): Total number of elements.
-    nn (int): Total number of nodes.
-    properties (dict): Material and physical constants, including:
-                       - h_conv, sigma_sb, T_amb, T_boiling, evc, Lev,
-                         cp_fluid, vareps, CM_coeff, CT_coeff, CP_coeff
-    F (array): Global force vector to be updated.
-
-    Returns:
-    array: Updated force vector with top-surface Neumann BCs applied.
-    """
-    # Extract physical constants
-    T_amb = properties["T_amb"]
-    T_boiling = properties["T_boiling"]
-    vareps = properties["vareps"]
-    evc = properties["evc"]
-    Lev = properties["Lev"]
-    cp_fluid = properties["cp_fluid"]
-    h_conv = properties["h_conv"]
-    sigma_sb = properties["sigma_sb"]
-
-    # Evaporation model coefficients (precomputed in SetupProperties)
-    invT_b = 1.0 / T_boiling
-    CM_coeff = properties["CM_coeff"]
-    CT_coeff = properties["CT_coeff"]
-    CP_coeff = properties["CP_coeff"]
-
-    # Mesh geometry
-    x, y = Level["node_coords"][0], Level["node_coords"][1]
-    cx, cy = Level["connect"][0], Level["connect"][1]
-    ne_x = jnp.size(cx, 0)
-    ne_y = jnp.size(cy, 0)
-    top_ne = ne - ne_x * ne_y
-    nn_x, nn_y = ne_x + 1, ne_y + 1
-
-    # Coordinates of a representative top-surface element
-    coords = jnp.stack([x[cx[0, :]], y[cy[0, :]]], axis=1)
-
-    # Precompute shape functions and quadrature weights
-    N, _, wq = computeQuad2dFemShapeFunctions_jax(coords)
-
-    def calcCR(i):
-        """
-        Compute the integrated heat flux vector for a single top-surface element.
-        """
-        _, _, _, idx = convert2XYZ(i, ne_x, ne_y, nn_x, nn_y)
-
-        # Interpolate nodal temperatures to quadrature points
-        Tq = jnp.matmul(N, LevelT0[idx[4:]])
-        Tq = jnp.minimum(Tq, T_boiling + 1000)
-
-        invT = 1.0 / Tq
-
-        # Energy required for phase change and heating of vapor
-        # Includes latent heat and sensible heat from ambient to surface temperature
-        E_pv = Lev + cp_fluid * (Tq - T_amb)  # J/kg
-
-        # Molecular motion term from kinetic theory
-        # Proportional to sqrt(m / (2πRT)), affects evaporation rate
-        MolMot = jnp.sqrt(CM_coeff * invT)  # s/m
-
-        # Evaporative heat loss (W/m²)
-        S = evc * CP_coeff * jnp.exp(-CT_coeff * (invT - invT_b)) * MolMot * E_pv
-
-        # Total heat flux (positive into the domain)
-        q_flux = h_conv * (T_amb - Tq) + sigma_sb * vareps * (T_amb**4 - Tq**4) - S
-
-        # Convert from W/m² to W/mm²
-        q_flux *= 1e-6
-
-        # Integrate over the element
-        return jnp.matmul(N.T, q_flux * wq.reshape(-1)), idx[4:]
-
-    # Vectorized computation over all top-surface elements
-    aT, aidx = jax.vmap(calcCR)(jnp.arange(top_ne, ne))
-
-    # Assemble into global force vector
-    NeumannBC = jnp.bincount(aidx.reshape(-1), aT.reshape(-1), length=nn)
-
-    return F + NeumannBC
-
-
-@jax.jit
-def moveEverything(v, vstart, Levels, move_v, LInterp, L1L2Eratio, L2L3Eratio, height):
-    """
-    Move the multilevel mesh system in response to laser motion.
-
-    This function updates node coordinates, interpolates temperature and subgrid
-    fields, and recalculates shape functions for Levels 1-3 based on the laser's
-    movement vector.
-
-    Parameters:
-    v (array): Current laser position.
-    vstart (array): Initial laser position.
-    Levels (dict): Multilevel mesh and field data.
-    move_v (array): Previous movement vector (used for scaling).
-    LInterp (list): Interpolation matrices between levels.
-    L1L2Eratio (list): Element ratio from Level 1 to Level 2 in [x, y, z].
-    L2L3Eratio (list): Element ratio from Level 2 to Level 3 in [x, y, z].
-    height (float): Layer height in mm.
-
-    Returns:
-    tuple:
-        - Levels (dict): Updated mesh and field data.
-        - Shapes (list): Updated shape function data.
-        - LInterp (list): Updated interpolation matrices.
-        - move_v (array): Updated movement vector.
-    """
-    # --- Move Level 3 (finest) ---
-    vtot = v - vstart
-    v_L3_constrained = jit_constrain_v(vtot, Levels[3])
-
-    # Move mesh and update coordinates
-    new_coords_L3, _ = move_fine_mesh(
-        Levels[3]["init_node_coors"], Levels[2]["h"], v_L3_constrained
-    )
-    Levels[3] = update_overlap_nodes_coords(
-        Levels[3], v_L3_constrained, Levels[2]["h"], [1, 1, 1]
-    )
-
-    # Interpolate temperature and subgrid fields
-    Levels[3]["T0"] = interpolatePoints(Levels[1], Levels[1]["T0"], new_coords_L3)
-    Tprime_L2 = interpolatePoints(Levels[2], Levels[2]["Tprime0"], new_coords_L3)
-    Tprime_L3 = interpolatePoints(Levels[3], Levels[3]["Tprime0"], new_coords_L3)
-    Levels[3]["T0"] += Tprime_L2 + Tprime_L3
-    Levels[3]["Tprime0"] = Tprime_L3
-    Levels[3]["node_coords"] = new_coords_L3
-
-    # --- Move Level 2 ---
-    v_L2_constrained = jit_constrain_v(vtot, Levels[2])
-    h_L1 = Levels[1]["h"][:2]
-    h_L1.append(jnp.array(height))
-    new_coords_L2, move_v = move_fine_mesh(
-        Levels[2]["init_node_coors"], h_L1, v_L2_constrained
-    )
-    move_v = [move_v[i] * L1L2Eratio[i] for i in range(3)]
-
-    Levels[2] = update_overlap_nodes_coords_L1L2(
-        Levels[2], v_L2_constrained, Levels[1]["h"], height
-    )
-
-    # Interpolate temperature and subgrid fields
-    Levels[2]["T0"] = interpolatePoints(Levels[1], Levels[1]["T0"], new_coords_L2)
-    Levels[2]["Tprime0"] = interpolatePoints(
-        Levels[2], Levels[2]["Tprime0"], new_coords_L2
-    )
-    Levels[2]["T0"] += Levels[2]["Tprime0"]
-    Levels[2]["node_coords"] = new_coords_L2
-
-    # Recompute shape functions and interpolation matrix
-    L2L1Shape = computeCoarseFineShapeFunctions(Levels[1], Levels[2])
-    LInterp[0] = interpolatePointsMatrix(Levels[1], new_coords_L2)
-
-    # --- Update Level 3 overlap nodes after Level 2 move ---
-    Levels[3]["overlapNodes"] = [
-        Levels[3]["overlapNodes"][i] - move_v[i] for i in range(3)
-    ]
-
-    # --- Update Level 0 (global) ---
-    Levels[0] = update_overlap_nodes_coords(
-        Levels[0], v_L3_constrained, Levels[2]["h"], L2L3Eratio
-    )
-    Levels[0] = update_overlap_nodes_coords_L2(
-        Levels[0],
-        v_L2_constrained,
-        [Levels[1]["h"][0], Levels[1]["h"][1], Levels[2]["h"][2]],
-        [L1L2Eratio[0] * L2L3Eratio[0], L1L2Eratio[1] * L2L3Eratio[1], L2L3Eratio[2]],
-    )
-
-    # Track z-direction movement only for Level 0
-    Levels[0]["overlapNodes"][2] -= move_v[2] * L2L3Eratio[2]
-    Levels[0]["overlapNodes_L2"][2] -= move_v[2] * L2L3Eratio[2]
-
-    # Update overlap indices
-    Levels[0]["idx"] = getOverlapRegion(
-        Levels[0]["overlapNodes"], Levels[0]["nodes"][0], Levels[0]["nodes"][1]
-    )
-    Levels[0]["idx_L2"] = getOverlapRegion(
-        Levels[0]["overlapNodes_L2"], Levels[0]["nodes"][0], Levels[0]["nodes"][1]
-    )
-
-    # Update subgrid state fields
-    Levels[2]["S1"] = Levels[2]["S1"].at[:].set(Levels[0]["S1"][Levels[0]["idx_L2"]])
-    Levels[3]["S1"] = Levels[3]["S1"].at[:].set(Levels[0]["S1"][Levels[0]["idx"]])
-    Levels[3]["S2"] = Levels[3]["S2"].at[:].set(Levels[0]["S2"][Levels[0]["idx"]])
-
-    # Recompute shape functions for updated meshes
-    L3L1Shape = computeCoarseFineShapeFunctions(Levels[1], Levels[3])
-    L3L2Shape = computeCoarseFineShapeFunctions(Levels[2], Levels[3])
-    LInterp[1] = interpolatePointsMatrix(Levels[2], new_coords_L3)
-
-    Shapes = [L2L1Shape, L3L1Shape, L3L2Shape]
-    return Levels, Shapes, LInterp, move_v
 
 
 @partial(jax.jit, static_argnames=["substrate"])
@@ -1885,403 +408,6 @@ def computeStateProperties(T, S1, properties, Level_nodes_substrate):
     return S1, S2, k, rhocp
 
 
-@partial(jax.jit, static_argnames=["ne_nn", "tmp_ne_nn", "substrate"])
-def stepGOMELTDwellTime(Levels, tmp_ne_nn, ne_nn, properties, dt, substrate):
-    """
-    Perform a dwell-time update for Level 1 in the GO-MELT framework.
-
-    This function updates the temperature field of Level 1 during laser dwell time,
-    assuming no motion of finer levels. It includes convection, radiation, and
-    evaporation boundary conditions, and enforces inactive nodes above the powder line.
-
-    Parameters:
-    Levels (list of dict): Multilevel simulation state. Only Levels[1] is updated.
-    tmp_ne_nn (tuple): Level 1 metadata.
-        - tmp_ne_nn[0] (int): Number of elements in Level 1.
-        - tmp_ne_nn[1] (int): Index above which nodes are inactive (above surface).
-    ne_nn (tuple): Neighbor and node connectivity data.
-        - ne_nn[2]: Connectivity for Level 1.
-    properties (dict): Material properties.
-    dt (float): Time step size (s).
-    substrate (tuple): Substrate information.
-        - substrate[1]: Number of substrate nodes (always solid).
-
-    Returns:
-    list of dict: Updated Levels list with modified temperature field in Levels[1].
-    """
-    L1 = Levels[1]
-    num_elements_L1 = tmp_ne_nn[0]
-    inactive_start_idx = tmp_ne_nn[1]
-
-    # Compute boundary conditions (includes convection, radiation, evaporation)
-    Fc = computeConvRadBC(L1, L1["T0"], num_elements_L1, ne_nn[2], properties, F=0)
-
-    # Compute temperature-dependent properties
-    _, _, k, rhocp = computeStateProperties(
-        L1["T0"], L1["S1"], properties, substrate[1]
-    )
-
-    # Solve temperature field using matrix-free FEM
-    T_new = solveMatrixFreeFE(
-        L1, ne_nn[2], num_elements_L1, k, rhocp, dt, L1["T0"], Fc, 0
-    )
-
-    # Enforce ambient temperature in inactive region above surface
-    T_new = substitute_Tbar(T_new, inactive_start_idx, properties["T_amb"])
-
-    # Apply boundary conditions and update Level 1 temperature
-    Levels[1]["T0"] = assignBCs(T_new, Levels)
-
-    return Levels
-
-
-@partial(jax.jit, static_argnames=["ne_nn"])
-def computeL1TprimeTerms_Part1(Levels, ne_nn, L3k, Shapes, L2k):
-    """
-    Compute the subgrid correction terms for Level 1 from Levels 2 and 3.
-
-    This function calculates the divergence of the subgrid heat fluxes from
-    Level 2 and Level 3 and projects them onto the Level 1 mesh. These terms
-    are used in the predictor step of the GO-MELT predictor-corrector scheme.
-
-    Parameters:
-    Levels (list of dict): Multilevel simulation state.
-        - Levels[2]["Tprime0"]: Subgrid temperature field from Level 2.
-        - Levels[3]["Tprime0"]: Subgrid temperature field from Level 3.
-    ne_nn (tuple): Mesh metadata.
-        - ne_nn[0]: Number of elements in Level 2.
-        - ne_nn[1]: Number of elements in Level 3.
-    L3k (array): Thermal conductivity field for Level 3.
-    Shapes (tuple): Shape function data for projection.
-        - Shapes[0][1]: ∇N₂ᵀ ∇N₁ mapping for Level 2 to Level 1.
-        - Shapes[0][2]: Projection matrix from Level 2 to Level 1.
-        - Shapes[1][1]: ∇N₃ᵀ ∇N₁ mapping for Level 3 to Level 1.
-        - Shapes[1][2]: Projection matrix from Level 3 to Level 1.
-    L2k (array): Thermal conductivity field for Level 2.
-
-    Returns:
-    array: Combined subgrid correction term projected onto Level 1.
-    """
-
-    # --- Level 3 subgrid flux divergence ---
-    coordsf = getSampleCoords(Levels[3])
-    Nf, dNdxf, wqf = computeQuad3dFemShapeFunctions_jax(coordsf)
-
-    _, _, _, idxf = convert2XYZ(
-        jnp.arange(ne_nn[1]),
-        Levels[3]["elements"][0],
-        Levels[3]["elements"][1],
-        Levels[3]["nodes"][0],
-        Levels[3]["nodes"][1],
-    )
-    _L3Tp0 = Levels[3]["Tprime0"][idxf]
-    L3kMean = jnp.matmul(Nf, L3k[idxf]).mean(axis=0)
-
-    # Compute ∇·(k ∇T') for Level 3
-    dL3Tp0dX = [multiply(L3kMean, (dNdxf[:, :, i] @ _L3Tp0)) for i in range(3)]
-    _wqf = wqf[None, None, :]
-    _dL3L1dX = Shapes[1][1]
-
-    _1 = sum(
-        multiply(multiply(-_dL3L1dX[i], dL3Tp0dX[i].T[:, :, None]), _wqf).sum(axis=2)
-        for i in range(3)
-    )
-
-    # --- Level 2 subgrid flux divergence ---
-    coordsm = getSampleCoords(Levels[2])
-    Nm, dNdxm, wqm = computeQuad3dFemShapeFunctions_jax(coordsm)
-
-    _, _, _, idxm = convert2XYZ(
-        jnp.arange(ne_nn[0]),
-        Levels[2]["elements"][0],
-        Levels[2]["elements"][1],
-        Levels[2]["nodes"][0],
-        Levels[2]["nodes"][1],
-    )
-    _L2Tp0 = Levels[2]["Tprime0"][idxm]
-    L2kMean = jnp.matmul(Nm, L2k[idxm]).mean(axis=0)
-
-    # Compute ∇·(k ∇T') for Level 2
-    dL2Tp0dX = [multiply(L2kMean, (dNdxm[:, :, i] @ _L2Tp0)) for i in range(3)]
-    _wqm = wqm[None, None, :]
-    _dL2L1dX = Shapes[0][1]
-
-    _2 = sum(
-        multiply(multiply(-_dL2L1dX[i], dL2Tp0dX[i].T[:, :, None]), _wqm).sum(axis=2)
-        for i in range(3)
-    )
-
-    # Project both subgrid terms to Level 1
-    return Shapes[1][2] @ _1.reshape(-1) + Shapes[0][2] @ _2.reshape(-1)
-
-
-@partial(jax.jit, static_argnames=["ne_nn", "tmp_ne_nn"])
-def computeL1Temperature(
-    Levels, ne_nn, tmp_ne_nn, L1F, L1V, L1k, L1rhocp, dt, properties
-):
-    """
-    Compute the updated temperature field for Level 1.
-
-    This function performs the temperature update for Level 1 using a matrix-free
-    finite element solver. It includes source terms, subgrid corrections, and
-    enforces boundary and ambient conditions.
-
-    Parameters:
-    Levels (list of dict): Multilevel simulation state.
-        - Levels[1]["T0"]: Current temperature field for Level 1.
-    ne_nn (tuple): Mesh connectivity data.
-        - ne_nn[2]: Connectivity for Level 1.
-    tmp_ne_nn (tuple): Level 1 metadata.
-        - tmp_ne_nn[0]: Number of elements in Level 1.
-        - tmp_ne_nn[1]: Index above which nodes are inactive (above surface).
-    L1F (array): Heat source vector for Level 1.
-    L1V (array): Subgrid correction vector for Level 1.
-    L1k (array): Thermal conductivity field for Level 1.
-    L1rhocp (array): Volumetric heat capacity field for Level 1.
-    dt (float): Time step size (s).
-    properties (dict): Material and process properties.
-        - properties["T_amb"]: Ambient temperature (K).
-
-    Returns:
-    array: Updated temperature field for Level 1.
-    """
-    # Solve temperature field using matrix-free FEM
-    L1T = solveMatrixFreeFE(
-        Levels[1], ne_nn[2], tmp_ne_nn[0], L1k, L1rhocp, dt, Levels[1]["T0"], L1F, L1V
-    )
-
-    # Apply ambient temperature to inactive region above surface
-    L1T = substitute_Tbar(L1T, tmp_ne_nn[1], properties["T_amb"])
-
-    # Enforce Dirichlet boundary conditions
-    FinalL1 = assignBCs(L1T, Levels)
-
-    return FinalL1
-
-
-@partial(jax.jit, static_argnames=["ne_nn"])
-def computeL2TprimeTerms_Part1(Levels, ne_nn, L3Tprime0, L3k, Shapes):
-    """
-    Compute the subgrid correction term for Level 2 from Level 3.
-
-    This function calculates the divergence of the subgrid heat flux from Level 3
-    and projects it onto the Level 2 mesh. This term is used in the predictor step
-    of the GO-MELT predictor-corrector scheme.
-
-    Parameters:
-    Levels (list of dict): Multilevel simulation state.
-        - Levels[3]["elements"], ["nodes"]: Mesh structure for Level 3.
-    ne_nn (tuple): Mesh metadata.
-        - ne_nn[1]: Number of elements in Level 3.
-    L3Tprime0 (array): Subgrid temperature field from Level 3.
-    L3k (array): Thermal conductivity field for Level 3.
-    Shapes (tuple): Shape function data.
-        - Shapes[2][1]: ∇N₃ᵀ ∇N₂ mapping for Level 3 to Level 2.
-        - Shapes[2][2]: Projection matrix from Level 3 to Level 2.
-
-    Returns:
-    array: Subgrid correction term projected onto Level 2.
-    """
-    # Get quadrature shape functions and weights for Level 3
-    coordsf = getSampleCoords(Levels[3])
-    Nf, dNdxf, wqf = computeQuad3dFemShapeFunctions_jax(coordsf)
-
-    # Get global node indices for all Level 3 elements
-    _, _, _, idxf = convert2XYZ(
-        jnp.arange(ne_nn[1]),
-        Levels[3]["elements"][0],
-        Levels[3]["elements"][1],
-        Levels[3]["nodes"][0],
-        Levels[3]["nodes"][1],
-    )
-
-    # Extract subgrid temperature and conductivity
-    _L3Tp0 = L3Tprime0[idxf]
-    L3kMean = jnp.matmul(Nf, L3k[idxf]).mean(axis=0)
-
-    # Compute ∇·(k ∇T') for Level 3
-    dL3Tp0dx = multiply(L3kMean, (dNdxf[:, :, 0] @ _L3Tp0))
-    dL3Tp0dy = multiply(L3kMean, (dNdxf[:, :, 1] @ _L3Tp0))
-    dL3Tp0dz = multiply(L3kMean, (dNdxf[:, :, 2] @ _L3Tp0))
-
-    # Weighted divergence projection to Level 2
-    _1 = multiply(
-        multiply(-Shapes[2][1][0], dL3Tp0dx.T[:, :, None]), wqf[None, None, :]
-    ).sum(axis=2)
-    _1 += multiply(
-        multiply(-Shapes[2][1][1], dL3Tp0dy.T[:, :, None]), wqf[None, None, :]
-    ).sum(axis=2)
-    _1 += multiply(
-        multiply(-Shapes[2][1][2], dL3Tp0dz.T[:, :, None]), wqf[None, None, :]
-    ).sum(axis=2)
-
-    # Project to global Level 2 vector
-    return Shapes[2][2] @ _1.reshape(-1)
-
-
-@partial(jax.jit, static_argnames=["ne_nn"])
-def computeL2Temperature(
-    L1T, L1L2Interp, Levels, ne_nn, L2T0, L2F, L2V, L2k, L2rhocp, dt
-):
-    """
-    Compute the updated temperature field for Level 2.
-
-    This function solves the meso-scale (Level 2) temperature field using
-    matrix-free FEM, incorporating subgrid corrections and boundary conditions
-    interpolated from the coarse-scale (Level 1) solution.
-
-    Parameters:
-    L1T (array): Temperature field from Level 1.
-    L1L2Interp (array): Interpolation matrix from Level 1 to Level 2 boundary.
-    Levels (list of dict): Multilevel simulation state.
-        - Levels[2]: Contains Level 2 mesh and boundary condition metadata.
-    ne_nn (tuple): Mesh metadata.
-        - ne_nn[0]: Number of elements in Level 2.
-        - ne_nn[3]: Connectivity for Level 2.
-    L2T0 (array): Initial temperature field for Level 2.
-    L2F (array): Heat source vector for Level 2.
-    L2V (array): Subgrid correction vector for Level 2.
-    L2k (array): Thermal conductivity field for Level 2.
-    L2rhocp (array): Volumetric heat capacity field for Level 2.
-    dt (float): Time step size (s).
-
-    Returns:
-    array: Updated temperature field for Level 2.
-    """
-    # Interpolate Level 1 temperature to Level 2 boundary
-    TfAll = interpolate_w_matrix(L1L2Interp, L1T)
-
-    # Solve Level 2 temperature using matrix-free FEM
-    L2T = solveMatrixFreeFE(
-        Levels[2], ne_nn[3], ne_nn[0], L2k, L2rhocp, dt, L2T0, L2F, L2V
-    )
-
-    # Apply Dirichlet boundary conditions from interpolated Level 1 solution
-    FinalL2 = assignBCsFine(L2T, TfAll, Levels[2]["BC"])
-
-    return FinalL2
-
-
-@partial(jax.jit, static_argnames=["ne_nn"])
-def computeSolutions_L3(
-    FinalL2, L2L3Interp, Levels, ne_nn, L3T0, L3F, L3k, L3rhocp, dt
-):
-    """
-    Compute the updated temperature field for Level 3.
-
-    This function solves the fine-scale (Level 3) temperature field using
-    matrix-free FEM, applying Dirichlet boundary conditions interpolated
-    from the meso-scale (Level 2) solution.
-
-    Parameters:
-    FinalL2 (array): Updated temperature field from Level 2.
-    L2L3Interp (array): Interpolation matrix from Level 2 to Level 3 boundary.
-    Levels (list of dict): Multilevel simulation state.
-        - Levels[3]: Contains Level 3 mesh and boundary condition metadata.
-    ne_nn (tuple): Mesh metadata.
-        - ne_nn[1]: Number of elements in Level 3.
-        - ne_nn[4]: Connectivity for Level 3.
-    L3T0 (array): Initial temperature field for Level 3.
-    L3F (array): Heat source vector for Level 3.
-    L3k (array): Thermal conductivity field for Level 3.
-    L3rhocp (array): Volumetric heat capacity field for Level 3.
-    dt (float): Time step size (s).
-
-    Returns:
-    array: Updated temperature field for Level 3.
-    """
-    # Interpolate Level 2 temperature to Level 3 boundary
-    TfAll = interpolate_w_matrix(L2L3Interp, FinalL2)
-
-    # Solve Level 3 temperature using matrix-free FEM
-    FinalL3 = solveMatrixFreeFE(
-        Levels[3], ne_nn[4], ne_nn[1], L3k, L3rhocp, dt, L3T0, L3F, 0
-    )
-
-    # Apply Dirichlet boundary conditions from interpolated Level 2 solution
-    FinalL3 = assignBCsFine(FinalL3, TfAll, Levels[3]["BC"])
-
-    return FinalL3
-
-
-@partial(jax.jit, static_argnames=["ne_nn"])
-def computeL1TprimeTerms_Part2(
-    Levels, ne_nn, L3Tp, L2Tp, L3rhocp, L2rhocp, dt, Shapes, Vcu
-):
-    """
-    Compute the time derivative subgrid correction terms for Level 1.
-
-    This function calculates the contribution of the time derivative of the
-    subgrid temperature fields from Levels 2 and 3, and projects them onto
-    the Level 1 mesh. These terms are added to the Level 1 correction vector.
-
-    Parameters:
-    Levels (list of dict): Multilevel simulation state.
-        - Levels[2]["Tprime0"], Levels[3]["Tprime0"]: Previous subgrid fields.
-    ne_nn (tuple): Mesh metadata.
-        - ne_nn[0]: Number of elements in Level 2.
-        - ne_nn[1]: Number of elements in Level 3.
-    L3Tp (array): Updated subgrid temperature field for Level 3.
-    L2Tp (array): Updated subgrid temperature field for Level 2.
-    L3rhocp (array): Volumetric heat capacity for Level 3.
-    L2rhocp (array): Volumetric heat capacity for Level 2.
-    dt (float): Time step size (s).
-    Shapes (tuple): Shape function data.
-        - Shapes[0][0]: N₂ᵀ N₁ mapping for Level 2 to Level 1.
-        - Shapes[0][2]: Projection matrix from Level 2 to Level 1.
-        - Shapes[1][0]: N₃ᵀ N₁ mapping for Level 3 to Level 1.
-        - Shapes[1][2]: Projection matrix from Level 3 to Level 1.
-    Vcu (array): Level 1 correction vector to be updated.
-
-    Returns:
-    array: Updated Level 1 correction vector including time derivative terms.
-    """
-    # Compute change in subgrid temperature fields
-    L3Tp_new = L3Tp - Levels[3]["Tprime0"]
-    L2Tp_new = L2Tp - Levels[2]["Tprime0"]
-
-    # --- Level 3 contribution ---
-    coordsf = getSampleCoords(Levels[3])
-    Nf, _, wqf = computeQuad3dFemShapeFunctions_jax(coordsf)
-
-    _, _, _, idxf = convert2XYZ(
-        jnp.arange(ne_nn[1]),
-        Levels[3]["elements"][0],
-        Levels[3]["elements"][1],
-        Levels[3]["nodes"][0],
-        Levels[3]["nodes"][1],
-    )
-
-    # Integrate ρcp * T' over Level 3 elements
-    _L3Tp = multiply(Nf @ L3Tp_new[idxf], jnp.matmul(Nf, L3rhocp[idxf]).mean(axis=0))
-    _1 = multiply(
-        multiply(-Shapes[1][0], _L3Tp.T[:, :, None]), (1 / dt) * wqf[None, None, :]
-    ).sum(axis=2)
-
-    # --- Level 2 contribution ---
-    coordsm = getSampleCoords(Levels[2])
-    Nm, _, wqm = computeQuad3dFemShapeFunctions_jax(coordsm)
-
-    _, _, _, idxm = convert2XYZ(
-        jnp.arange(ne_nn[0]),
-        Levels[2]["elements"][0],
-        Levels[2]["elements"][1],
-        Levels[2]["nodes"][0],
-        Levels[2]["nodes"][1],
-    )
-
-    # Integrate ρcp * T' over Level 2 elements
-    _L2Tp = multiply(Nm @ L2Tp_new[idxm], jnp.matmul(Nm, L2rhocp[idxm]).mean(axis=0))
-    _2 = multiply(
-        multiply(-Shapes[0][0], _L2Tp.T[:, :, None]), (1 / dt) * wqm[None, None, :]
-    ).sum(axis=2)
-
-    # Project both contributions to Level 1 and update correction vector
-    Vcu += Shapes[1][2] @ _1.reshape(-1) + Shapes[0][2] @ _2.reshape(-1)
-
-    return Vcu
-
-
 def getSampleCoords(Level):
     """
     Extracts the nodal coordinates of the first element in the mesh.
@@ -2296,61 +422,6 @@ def getSampleCoords(Level):
     y = Level["node_coords"][1][Level["connect"][1][0, :]].reshape(-1, 1)
     z = Level["node_coords"][2][Level["connect"][2][0, :]].reshape(-1, 1)
     return jnp.concatenate([x, y, z], axis=1)
-
-
-@partial(jax.jit, static_argnames=["ne_nn"])
-def computeL2TprimeTerms_Part2(Levels, ne_nn, L3Tp, L3Tp0, L3rhocp, dt, Shapes, L2V):
-    """
-    Compute the time derivative subgrid correction term for Level 2 from Level 3.
-
-    This function calculates the contribution of the time derivative of the
-    subgrid temperature field from Level 3 and projects it onto the Level 2 mesh.
-
-    Parameters:
-    Levels (list of dict): Multilevel simulation state.
-        - Levels[3]["Tprime0"]: Previous subgrid temperature field.
-    ne_nn (tuple): Mesh metadata.
-        - ne_nn[1]: Number of elements in Level 3.
-    L3Tp (array): Updated subgrid temperature field for Level 3.
-    L3Tp0 (array): Previous subgrid temperature field for Level 3.
-    L3rhocp (array): Volumetric heat capacity for Level 3.
-    dt (float): Time step size (s).
-    Shapes (tuple): Shape function data.
-        - Shapes[2][0]: N₃ᵀ N₂ mapping for Level 3 to Level 2.
-        - Shapes[2][2]: Projection matrix from Level 3 to Level 2.
-    L2V (array): Level 2 correction vector to be updated.
-
-    Returns:
-    array: Updated Level 2 correction vector including time derivative terms.
-    """
-    # Compute change in subgrid temperature field
-    L3Tp_new = L3Tp - L3Tp0
-
-    # Get shape functions and weights for Level 3
-    coordsf = getSampleCoords(Levels[3])
-    Nf, _, wqf = computeQuad3dFemShapeFunctions_jax(coordsf)
-
-    # Get global node indices for Level 3 elements
-    _, _, _, idxf = convert2XYZ(
-        jnp.arange(ne_nn[1]),
-        Levels[3]["elements"][0],
-        Levels[3]["elements"][1],
-        Levels[3]["nodes"][0],
-        Levels[3]["nodes"][1],
-    )
-
-    # Integrate ρcp * T' over Level 3 elements
-    _L3Tp = multiply(Nf @ L3Tp_new[idxf], jnp.matmul(Nf, L3rhocp[idxf]).mean(axis=0))
-
-    # Compute weighted projection to Level 2
-    _data2 = multiply(
-        multiply(-Shapes[2][0], _L3Tp.T[:, :, None]), (1 / dt) * wqf[None, None, :]
-    ).sum(axis=2)
-
-    # Update Level 2 correction vector
-    L2V += Shapes[2][2] @ _data2.reshape(-1)
-
-    return L2V
 
 
 def printLevelMaxMin(Ls, Lnames):
@@ -2384,34 +455,6 @@ def printLevelMaxMin(Ls, Lnames):
     if flag:
         sys.exit(1)
     print("")
-
-
-def saveResults(Levels, Nonmesh, savenum):
-    """
-    Save temperature results for Levels 1-3 based on save frequency and flags.
-    """
-    if Nonmesh["output_files"] == 1:
-        if savenum == 1 or (
-            np.mod(savenum, Nonmesh["Level1_record_step"]) == 1
-            or Nonmesh["Level1_record_step"] == 1
-        ):
-            saveResult(Levels[1], "Level1_", savenum, Nonmesh["save_path"], 2e-3)
-            # saveState(Levels[0], "Level0_", savenum, Nonmesh["save_path"], 0)
-
-        saveResult(Levels[2], "Level2_", savenum, Nonmesh["save_path"], 1e-3)
-        saveResult(Levels[3], "Level3_", savenum, Nonmesh["save_path"], 0)
-        print(f"Saved Levels_{savenum:08}")
-
-
-def saveResultsFinal(Levels, Nonmesh):
-    """
-    Save final temperature results for Levels 1-3.
-    """
-    if Nonmesh["output_files"] == 1:
-        saveFinalResult(Levels[1], "Level1_", Nonmesh["save_path"], 2e-3)
-        saveFinalResult(Levels[2], "Level2_", Nonmesh["save_path"], 1e-3)
-        saveFinalResult(Levels[3], "Level3_", Nonmesh["save_path"], 0)
-        print("Saved Final Results")
 
 
 def melting_temp(temps, delt_T, T_melt, accum_time, idx):
